+++ TM4C123_GPIO_LCD1602/xDriver_MCU/Driver_Source/SYSTICK/SYSTICK_Driver/SYSTICK_Intrinsics/SYSTICK_FreeCount.c
@@ -20,7 +20,7 @@ uint32_t SysTick__u32GetFreeCount(void)
 
 void SysTick__vClearFreeCount(void)
 {
-    SysTick_u32FreeCount=0;
+    SysTick_u32FreeCount=(uint32_t)0;
 }
 
 void SysTick__vSetFreeCount(uint32_t u32Count)
@@ -36,7 +36,7 @@ uint32_t SysTick__u32GetFreeCountOv(void)
 
 void SysTick__vClearFreeCountOv(void)
 {
-    SysTick_u32FreeCountOv=0;
+    SysTick_u32FreeCountOv=(uint32_t)0;
 }
 
 void SysTick__vSetFreeCountOv(uint32_t u32Count)
